% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipScaleFactors.R
\name{clipScaleFactors}
\alias{clipScaleFactors}
\title{Calculate scale factors for CLIP-seq data}
\usage{
clipScaleFactors(
  wigs,
  data_table,
  sdn = 5,
  crossnormalize = T,
  plot = T,
  bg_cut = 5,
  breakpoint = "kmeans",
  adjust = 2
)
}
\arguments{
\item{wigs}{A list of IRanges rle coverage vectors.}

\item{data_table}{Vector metadata in data frame format, see package vignette
for details.}

\item{sdn}{Number of standard deviations for elliptical filtering, see
\code{\link{filter_elliptical}} for details.}

\item{crossnormalize}{Crossnormalize the control libraries?}

\item{plot}{Produce diagnostic plots?}

\item{bg_cut}{Cut-off for control library normalization.}
}
\value{
Returns an array of scale factors for the provided libraries.
}
\description{
Calculate scale factors for CLIP-seq data
}
\details{
This function computes scale factors for CLIP-seq data, under the
assumption that it follows a bi- or multi-modal distribution. For each
matched experiment, low count reads are first filtered using an elliptical
filter based on the read count standard deviation. Log2 ratios of read counts
per nucleotide are calculated, and a local minima is identified between the
two highest density maxima. Positions with a log2 ratio less than this minima
are then used to fit a scale factor for each experiment. Optionally using the
\code{crossnormalize} parameter, all libraries can be scaled based on scale
factors computed between control libraries to produce consistent values for
hypothesis testing. Note positions with less than \code{bg_cut} reads in
any library will be excluded.
}
\examples{

}
\seealso{
\code{\link{norclip}}, \code{\link{loadData}},
\code{\link{gm_scale_factors}}, \code{\link{filter_elliptical}}
\code{\link{clip_scat}}
}
