% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_elliptical.R
\name{filter_elliptical}
\alias{filter_elliptical}
\title{Construct elliptical filters for matched CLIP-seq experiments}
\usage{
filter_elliptical(exp, ctrl, sdn = 8)
}
\arguments{
\item{exp}{A (possibly IRanges rle) vector containing read-counts per
nucleotide in a crosslinked sample}

\item{ctrl}{A (possibly IRanges rle) vector containing read-counts per
nucleotide in a non-crosslinked control}

\item{sdn}{The number of standard deviations to use in constructing the
elliptical filter}
}
\value{
Returns a (possibly IRanges rle) elliptical filter mask.
}
\description{
This function constructs an elliptical filter centered on the origin for two
libraries based on the per nucleotide read counts.
The filter is defined as \code{((exp)^2)/a + ((ctrl)^2)/b = 1}, where
\code{a} and \code{b} are the means of each coverage vector plus \code{sdn}
standard deviations. Positions within the ellipse return \code{FALSE}, while
those outside the ellipse return \code{TRUE}. Will return an IRanges rle
vector if the input vectors are rle vectors.
}
\examples{

}
