% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_minima_in_range.R
\name{find_minima_in_range}
\alias{find_minima_in_range}
\title{Find local minima in the empirical distribution of a vector.}
\usage{
find_minima_in_range(
  data,
  n = 1,
  range = c(min(data), max(data)),
  plot = FALSE,
  density_est = "empirical",
  adjust = 2,
  components = 2
)
}
\arguments{
\item{data}{A vector containing numeric data points.}

\item{n}{Number of local minima to find in the empirical distribution.}

\item{range}{Range in data within which to find minima.}

\item{plot}{Plot density and found minima?}
}
\value{
A vector containing n local minima ordered by
decreasing density.
}
\description{
Heuristic function to find local minima in the empirical distribution of a
vector of values within a certain range. Usually used with the output of
\code{\link{find_maxima}} to identify a local minimum. Minima are identified
as appropriate change points in a density curve fitted to the input data, and
are reported in order of increasing density in the case of multiple minima.
Will optionally plot the density and predicted minima over a histogram of the
input data.
}
\examples{

x <- c(rnorm(100), rnorm(100, mean=4))
maxima <- find_maxima(x)
find_minima_in_range(x, range=c(min(maxima), max(maxima)), plot=TRUE)

}
\seealso{
\code{\link{find_maxmima}}
}
