% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_separator_dbscan.R,
%   R/find_separator_kmeans.R
\name{find_separator_kmeans}
\alias{find_separator_kmeans}
\title{Find separation point between two clusters in a 1D vector}
\usage{
find_separator_kmeans(data, plot = TRUE)

find_separator_kmeans(data, plot = TRUE)
}
\arguments{
\item{data}{A 1-dimensional vector}

\item{plot}{Plot histogram?}
}
\value{
Returns value for separation point between clusters

Returns value for separation point between clusters
}
\description{
Find separation point between two clusters in a 1D vector

Find separation point between two clusters in a 1D vector
}
\details{
This function uses the \code{\link[Ckmeans.1d.dp]{Ckmeans.1d.dp}} package to
compute an optimal 2-component kmeans clustering on the given data, returning
the lowest value within the cluster with the highest mean to be used as a
separator in CLIP normalization.

This function uses the \code{\link[Ckmeans.1d.dp]{Ckmeans.1d.dp}} package to
compute an optimal 2-component kmeans clustering on the given data, returning
the lowest value within the cluster with the highest mean to be used as a
separator in CLIP normalization.
}
\examples{


}
\seealso{
\code{\link{norclip}}, \code{\link{clipScaleFactors}},
\code{\link{Ckmeans.1d.dp}}

\code{\link{norclip}}, \code{\link{clipScaleFactors}},
\code{\link{Ckmeans.1d.dp}}
}
