% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{A function for calculating the geometric mean of a vector}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{A vector containing}

\item{na.rm}{Remove undefined values?}

\item{zero.propagate}{Propagate zeros?}
}
\value{
gm_mean The geometric mean of a vector.
}
\description{
A function for calculating the geometric mean of a vector
}
\examples{

}
