% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norclip.R
\name{norclip}
\alias{norclip}
\title{Perform norclip normalization of CLIP-seq libraries}
\usage{
norclip(
  data_table,
  crossnormalize = T,
  sdn = 5,
  diagnostics = T,
  bg_cut = 5,
  naive = T,
  export_wigs = F,
  adjust = 2,
  breakpoint = "kmeans",
  replicons
)
}
\arguments{
\item{data_table}{A description table}

\item{crossnormalize}{Perform cross-normalization of replicates?}

\item{sdn}{Number of standard deviations for elliptical filtering}

\item{diagnostics}{Run diagnostics prior to normalization?}

\item{bg_cut}{Number of reads required in background library for a position
to be used in normalization}

\item{naive}{Additionally run "naive" normalization?}
}
\description{
Perform norclip normalization of CLIP-seq libraries
}
\examples{

}
