% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiagnostics.R
\name{runDiagnostics}
\alias{runDiagnostics}
\title{A function to generate diagnostic plots for CLIP-seq data.}
\usage{
runDiagnostics(
  wigs,
  data_table,
  sdn = 10,
  colramp = rev(rainbow(10, end = 4/6))
)
}
\arguments{
\item{wigs}{A list of IRanges rle coverage vectors.}

\item{data_table}{Vector metadata in data frame format, see package vignette
for details.}

\item{sdn}{Number of standard deviations for elliptical filtering, see
\code{\link{filter_elliptical}} for details.}

\item{colramp}{Color scheme for 2D density plots. Defaults to mimic the color
scheme of \code{\link{smoothScatter}}.}
}
\value{
No explicit return, will produce a correlation plot plus 3 plots per
experiment group: an unfiltered 2D density plot, an elliptically filtered 2D
density plot, and a density plot for the log2 ratio of experiment to control
library read counts.
}
\description{
This function produces various diagnostic plots to help the user understand
if their data is suitable for \code{\link{norclip}} analysis, and if the
default filtering values are appropriate. These include correlations between
libraries, unfiltered and filtered 2D density plots comparing crosslinked
libraries to uncrosslinked control libraries, and 1D histograms of the
resulting log2 ratios. See vignette for advice on interpreting these plots.
}
\examples{

}
\seealso{
\code{\link{norclip}}, \code{\link{loadData}},
\code{\link{filter_elliptical}}, \code{\link{clip_scat}}
}
