% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig2rle.R
\name{wig2rle}
\alias{wig2rle}
\title{Import wig files into rle coverage vectors.}
\usage{
wig2rle(forward_path, reverse_path)
}
\arguments{
\item{forward_path}{Path to forward strand wig file.}

\item{reverse_path}{Path to reverse strand wig file.}
}
\value{
Return a concatenated IRanges rle coverage vector.
}
\description{
Imports wiggle files for forward and reverse strands of a CLIP-seq
experiment. These are then converted to rle coverage vectors, and the
absolute value of the reverse strand is taken to guard against negative
values before concatenation.
}
\examples{

}
\seealso{
\code{\link{import}}, \code{\link{loadData}}
}
